Functional Commutation
After a lot of tribulation, the commutation algorithm is able to make a motor spin in a controlled fashion.
Right now, the code is still a bit hacky --- for example, it's based on force-feeding hall states to the motor without paying attention
to the hall states in the real world, so we'll have to take care of that to close the feedback loop. Additionally, it could use a whole
lot of clean-up and not all of the behavior (for example the CCW states) seem to be working.